education complete crud

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\EducationController;
 use App\Http\Controllers\ExperienceController;
 use App\Http\Controllers\ShowController;
 use Illuminate\Support\Facades\Route;
@@ -19,9 +20,16 @@ Route::get('/',[ShowController::class,'show']);
 
 
 
-
+// experience
 Route::get('admin.create_experience',[ExperienceController::class,'create_experience']);
-Route::post('store_experience',[ExperienceController::class,'store_experience']);
-Route::delete('destroy_experience/{id}',[ExperienceController::class,'destroy_experience']);
+Route::post('admin.store_experience',[ExperienceController::class,'store_experience']);
+Route::delete('admin.destroy_experience/{id}',[ExperienceController::class,'destroy_experience']);
 Route::get('admin.show_experience/{id}',[ExperienceController::class,"show_experience"]);
 Route::put('admin.update_experience/{id}',[ExperienceController::class,"update_experience"]);
+
+// education
+Route::get('admin.create_education',[EducationController::class,'create_education']);
+Route::post('admin.store_education',[EducationController::class,'store_education']);
+Route::delete('admin.destroy_education/{id}',[EducationController::class,'destroy_education']);
+Route::get('admin.show_education/{id}',[EducationController::class,"show_education"]);
+Route::put('admin.update_education/{id}',[EducationController::class,"update_education"]);
